+++ zocromas_zocprotos/src/http/mas_http_get.c
@@ -44,7 +44,7 @@ mas_http_make_out_head_get( mas_rcontrol_t * prcontrol, mas_http_t * http )
   if ( http && http->content )
   {
     MAS_LOG( "to make out" );
-    MAS_LOG( "SIZE %lu", mas_fileinfo_data_size( http->content ) );
+    MAS_LOG( "SIZE %lu", ( unsigned long ) mas_fileinfo_data_size( http->content ) );
 
     http = mas_http_make_out_header_simple( http, "Accept-Ranges", "bytes" );
     /* {                                                        */
